update securityContext in the helm charts.
This is broken across a commit per chart. To make it simpler to review.

These changes implement:

`podSecurityContext` the securityContext of defined at the pod level.
and
`securityContext` which is defined at the container level.

This set of changes allows for the definition of securityContext at
either level.

I've incorporated existing securityContext in a handler to define the
defaults if no additional securityContext is provided.

For hubble-ui there is a "securityContext.enabled" flag that specifies a
static value based on whether that is true. I've kept that functionality
and will append this value when specified in addition to any context
that is added to the values file.

This change also enables the podSecurityContext to be defined in a
global value.

Signed-off-by: Duffie Cooley <dcooley@isovalent.com>